+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Unreleased
+## 🐛 Bugfixes
+- Fixed some bugs in CWL Runner ramping up its conformance from 126/378 to 130/378
+- Improved Array support
+- Support cwl.output.json handling
+- Support globs
 
 # v0.3.0
 ## 🚀 Features