+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v0.4.0
 ## 🚀 Features
 - Added `--no-defaults` flag to tool create which can be handy when using passwords as inputs